+++ src-ui/app/views/dashboard-view/views/power-automations-view/power-automations-view.component.ts
@@ -1,13 +1,21 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { noop } from '../../../../utils/animations';
+import { firstValueFrom } from 'rxjs';
+import { ActivatedRoute } from '@angular/router';
 
+type PowerAutomationsTab = 'CONTROLLERS_AND_TRACKERS' | 'BASE_STATIONS' | 'WINDOWS_POWER_POLICY';
 @Component({
   selector: 'app-power-automations-view',
   templateUrl: './power-automations-view.component.html',
   styleUrls: ['./power-automations-view.component.scss'],
   animations: [noop()],
 })
-export class PowerAutomationsViewComponent {
-  activeTab: 'CONTROLLERS_AND_TRACKERS' | 'BASE_STATIONS' | 'WINDOWS_POWER_POLICY' =
-    'CONTROLLERS_AND_TRACKERS';
+export class PowerAutomationsViewComponent implements OnInit {
+  activeTab: PowerAutomationsTab = 'CONTROLLERS_AND_TRACKERS';
+
+  constructor(private activatedRoute: ActivatedRoute) {}
+  async ngOnInit() {
+    const fragment = await firstValueFrom(this.activatedRoute.fragment);
+    if (fragment) this.activeTab = fragment as PowerAutomationsTab;
+  }
 }